backend/requirements.txt
requests
fastapi
uvicorn
python-dotenv
pydantic
confluent-kafka
redis[hiredis]
fastapi-csrf-protect
slowapi
authlib
bcrypt
pyotp
boto3
sqlalchemy
pymongo
python-logging-loki
